Organizing the code in folders. The generated js code will be now be outputed to the dist folder.
