+++ .env.example
@@ -1,5 +1,5 @@
 DB_USER=dev
-DB_PASSWORD=dev
+DB_PASSWORD=dev123
 DB_DATABASE=lovelystay_test
 DB_HOST=postgres
