+++ .gitignore
@@ -104,3 +104,4 @@ dist
 .tern-port
 
 postgres/data/*
+dist/*
