+++ Dockerfile
@@ -2,7 +2,7 @@ FROM node:10
 
 WORKDIR /app
 
-COPY ./* /app
+COPY ./src /app
 
 RUN npm i
