+++ package.json
@@ -2,9 +2,9 @@
   "name": "pg-promise-test",
   "version": "1.0.0",
   "description": "pg-promise-github-test",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
-    "test": "node ./node_modules/ts-node/dist/bin.js server.ts"
+    "test": "node ./node_modules/ts-node/dist/bin.js src/index.ts"
   },
   "keywords": [
     "github",
@@ -13,7 +13,6 @@
   "author": "João Cristóvão",
   "license": "ISC",
   "dependencies": {
-    "@types/bluebird": "^3.5.2",
     "@types/node": "^7.0.12",
     "@types/ramda": "0.0.6",
     "@types/request-promise": "^4.1.33",
@@ -28,10 +27,11 @@
     "request-promise": "^4.2.0",
     "slugify": "^1.6.5",
     "ts-node": "^3.0.2",
-    "typescript": "^2.2.2",
+    "typescript": "^4.5.5",
     "winston": "^2.3.1"
   },
   "devDependencies": {
+    "@types/bluebird": "^3.5.36",
     "@types/command-line-args": "^5.2.0",
     "@types/command-line-usage": "^5.0.2",
     "@types/pg-promise": "0.0.28"
